Refactor adapter files and update include paths in perception module
